backend/requirements.txt
fastAPI>=0.110.0
uvicorn>=0.28.0
PyJWT>=2.3.0
python-decouple>=3.8
jinja2>=3.1.3
pydantic>=1.9.0
requests>=2.26.0
pymongo>=3.12.0
python-multipart>=0.0.9
pydantic[email]>=1.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2_cffi>=21.1.0
pytz>=2022.1
plotly>=5.3.1
httpx>=0.21.0
pytest>=7.0.0